requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
streamlit==1.41.1
requests==2.32.3
kafka-python==2.0.2
pydantic==2.10.3
pandas==2.2.3
redis==5.2.1
scikit-learn==1.5.2
deltalake==0.20.2
pyarrow==18.1.0
